% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{f7Radio}
\alias{f7Radio}
\title{Modified Framework7 radio input}
\usage{
f7Radio(inputId, label, choices = NULL, selected = NULL)
}
\arguments{
\item{inputId}{Radio input id.}

\item{label}{Radio label}

\item{choices}{List of choices. Must be a nested list.}

\item{selected}{Selected element. NULL by default.}
}
\description{
\code{f7Radio} creates a radio button input.
}
